Added more work towards the xml file signature
added the the class to enclusure the signature node allowing references to be dynamically added to the signature node and with a get note option which returns a xmldom compatible node.
also added a method to encapsulate the canonicalizer into a function returning a promise instead of working with callbacks.